requirements.txt
minigrid==2.3.1
torch==2.2.0
tqdm>=4.66.0
PyYAML>=6.0
matplotlib>=3.8.0
